refactor(desktop): move app shell components to app/components
