+++ apps/desktop/src/app/components/ErrorBoundary/index.tsx
