+++ apps/desktop/src/main.tsx
@@ -1,7 +1,7 @@
 import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 import { App } from './App';
-import { ErrorBoundary } from './components/ErrorBoundary';
+import { ErrorBoundary } from './app/components/ErrorBoundary';
 import { bootstrapTheme } from './shared/lib/theme';
 import { migrateLegacyStorageKeys } from './shared/lib/storage-keys';
 import './styles.css';
